refactor: Exception 추가
